app/requirements.txt
aiohttp==3.7.4
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==22.1.0
beautifulsoup4==4.9.1
breadability==0.1.20
chardet==3.0.4
charset-normalizer==2.1.1
clean-text==0.6.0
click==8.1.3
colorama==0.4.5
Django==4.1.2
docopt==0.6.2
emoji==1.7.0
fake-useragent==0.1.11
frozenlist==1.3.1
ftfy==6.1.1
gunicorn==20.1.0
html2text==2020.1.16
idna==3.4
joblib==1.2.0
lxml==4.6.5
multidict==6.0.2
nest-asyncio==1.5.6
nltk==3.7
numpy==1.23.4
pycountry==22.3.5
python-dotenv==0.21.0
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.3
search-engine-parser==0.6.7
soupsieve==2.3.2.post1
sqlparse==0.4.3
sumy==0.10.0
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.4
Unidecode==1.3.6
urllib3==1.26.12
wcwidth==0.2.5
whitenoise==6.2.0
yarl==1.8.1
